requirements.txt
streamlit
langchain
sentence-transformers
langchain-groq
langchain-community
pyodbc
faiss-cpu